+++ browser-extension/src/core/balance/non-current-assets.ts
@@ -1,6 +1,4 @@
 import { computed } from 'vue';
-import { sumMaterialAmountPrice } from '@src/infrastructure/fio/cx';
-import { sitesStore } from '@src/infrastructure/prun-api/data/sites';
 import {
   partnerNonCurrentConditions,
   selfNonCurrentConditions,
@@ -10,30 +8,7 @@ import {
   sumMaterialsShipment,
   sumMaterialsPickup,
 } from '@src/core/balance/contract-conditions';
-import { sumMapValues } from '@src/core/balance/utils';
-import { getPlanetNameFromAddress } from '@src/infrastructure/prun-api/data/addresses';
-
-type AddressableId = string;
-
-const buildings = computed(() => {
-  const buildings = new Map<AddressableId, number>();
-  // Buildings
-  for (const site of sitesStore.all.value) {
-    let value = 0;
-    for (const building of site.platforms) {
-      value += sumMaterialAmountPrice(building.reclaimableMaterials);
-    }
-    if (value === 0) {
-      continue;
-    }
-
-    const name = getPlanetNameFromAddress(site.address)!;
-    buildings.set(name, (buildings.get(name) ?? 0) + value);
-  }
-  return buildings;
-});
-
-const buildingsTotal = computed(() => sumMapValues(buildings.value));
+import { buildingsTotal, currentBuildingValue } from '@src/core/balance/buildings';
 
 const bases = computed(() => 1);
 
@@ -59,7 +34,7 @@ const total = computed(() => {
 });
 
 export const nonCurrentAssets = {
-  buildings,
+  buildings: currentBuildingValue,
   buildingsTotal,
   bases,
   accountsReceivable,